requirements.txt
requests==2.31.0
playwright==1.41.0
pandas==2.2.0
openai==1.10.0
python-dotenv==1.0.1
pydantic==2.6.4
streamlit==1.32.0
plotly==5.19.0
schedule==1.2.1
watchdog==4.0.0
altair<5
